requirements.txt
legofy
psutil
opencv_python
Pybooru
python_telegram_bot
uptime
numpy
requests[socks]
matplotlib
zalgo_text
gTTS
Pillow
google_api_python_client
scikit_learn
PyYAML
beautifulsoup4
lxml